Module sublock::atomlock

A variant of RwLock based on dynamic checks (comparable to RefCell). A variant of RwLock with sublocks that can be opened for reading iff the main lock is currently opened for reading, opened for writing iff the main lock is currently opened for writing.

Structs

Liveness

MainLock
A variant of RwLock with sublocks that can be opened for reading iff the main lock is currently opened for reading, opened for writing iff the main lock is currently opened for writing.

SubCell

WriteGuard

Type Definitions

LockResult

ReadGuard

Implementations